add haskell 1.3 release notes
from12to13.html is from
https://www.haskell.org/definition/from12to13.html

I have not found an 'authoritative' version of preludechanges.html
on haskell.org, but I did find files of the same name at
http://www.ki.informatik.uni-frankfurt.de/doc/html/Haskell1.3/preludechanges.html
and
https://web.archive.org/web/20050319043520/http://ceng.metu.edu.tr/docs/hugs/preludechanges.html
and these two URLs host the same content, so I am guessing they are accurate
mirrors.

For both files, I have fixed and removed hyperlinks so that all of the links now
work.